agents/schema-agent/gcp/cloud-functions/schema-updater/requirements.txt
functions-framework==3.*
google-cloud-storage
requests
graphql-core
sentence-transformers
faiss-cpu
numpy
tqdm